botlib: fix AngleDiff mistyped double return to float
evidence proves it is float, matching Q3A's be_ai_move.c:220 cognate
(same author, float AngleDiff). The real ELF binary inserts a
unmodified-return paths -- this pattern only makes sense if the
from a float value at x87 extended precision requires an explicit
truncation store to enforce correct float semantics, whereas a true
double would need no such truncation at all.

Also restructured the two early-return branches into compound
assignment + a single shared trailing return (result -= 360.0; ...;
return result;), matching Q3A's actual statement shape. This
restructuring alone had no measurable effect, but combined with the
type fix produced a full match.

  BotTravel_Grapple:  331 -> 321 insn_diffs (positive ripple, one of two call sites)

PE-oracle: AngleDiff already fully matched before this change (MSVC's
codegen never inserted a truncating store here regardless of the
declared type) and still does after -- unaffected either way. Full
810-function audit_dll.sh ripple check is byte-for-byte identical to
the committed baseline, including BotTravel_Grapple's pre-existing
789 byte_diffs (unrelated register-allocation churn, unchanged by
this fix). Zero PE regressions.

Source-commit: 27d5760413efd00eb578d9af277b7375d3eed4cb

+++ botlib/be_ai_move.h
@@ -18,7 +18,7 @@ bot_moveresult_t *__cdecl BotMoveToGoal(bot_moveresult_t *a1, bot_movestate_t *m
 extern int dword_1006295C;
 extern libvar_t *libvar_laserhook;
 
-double __cdecl AngleDiff(float ang1, float ang2);
+float __cdecl AngleDiff(float ang1, float ang2);
 void __cdecl BotAddToAvoidReach(bot_movestate_t *ms, int number, float avoidtime);
 int __cdecl BotCheckBarrierJump(bot_movestate_t *ms, float *dir, float speed);
 int __cdecl BotCheckBlocked(bot_movestate_t *ms, float *dir, bot_moveresult_t *moveresult);
